chore: refactor useChainlinkPrice to support multicall
